Bug fixed. Add "settings error".

+++ __init__.py
@@ -27,43 +27,56 @@ class RemoteComputerSkill(MycroftSkill):
     def __init__(self):
         super(RemoteComputerSkill, self).__init__(name="RemoteComputerSkill")
 
-    @intent_handler(IntentBuilder("TurnOnIntent").require("Turn")
-                    .require("Computer").require("On"))
+    @intent_handler(IntentBuilder("ComputerOnIntent").require("Computer")
+                    .require("On").optionally("Turn"))
     def handle_turn_on_intent(self, message):
         try:
             config = self.config_core.get("RemoteComputerSkill", {})
 
             if not config == {}:
                 mac_address = str(config.get("mac_address"))
+
             else:
                 mac_address = str(self.settings.get("mac_address"))
 
-            re_mac = "[0-9a-f]{2}([-:]?)[0-9a-f]{2}(\\1[0-9a-f]{2}){4}$"
-
-            if re.match(re_mac, mac_address.lower()):
-                if ':' in mac_address:
-                    mac_address.replace(':', '.')
-                elif '-' in mac_address:
-                    mac_address.replace('-', '.')
-
-                if self.ask_yesno("ask.confirmation",
-                                  {"word": "power on"}) == "no":
-                    self.speak_dialog("okay")
-                    return
-
-                send_magic_packet(mac_address)
-
-                self.speak_dialog("computer.on")
-            else:
-                self.speak_dialog("invalid", {"word": "mac"})
-                return
+            if not mac_address:
+                raise Exception("None found.")
 
         except Exception as e:
-            self.speak_dialog("connection.error")
+            self.speak_dialog("settings.error")
             self.log.error(e)
+            return
 
-    @intent_handler(IntentBuilder("TurnOffIntent").require("Turn")
-                    .require("Computer").require("Off"))
+        re_mac = "[0-9a-f]{2}([-:]?)[0-9a-f]{2}(\\1[0-9a-f]{2}){4}$"        
+
+        if re.match(re_mac, mac_address.lower()):
+            if ':' in mac_address:
+                mac_address.replace(':', '.')
+
+            elif '-' in mac_address:
+                mac_address.replace('-', '.')
+
+        else:
+            self.speak_dialog("invalid", {"word": "mac"})
+            return
+
+        prompt_response = self.ask_yesno("ask.confirmation",
+                                         {"word": "power on"})
+
+        if prompt_response == "yes":
+            try:
+                send_magic_packet(mac_address)
+                self.speak_dialog("computer.on")
+
+            except Exception as e:
+                self.speak_dialog("connection.error")
+                self.log.error(e)
+
+        elif prompt_response == "no":
+            self.speak_dialog("okay")
+
+    @intent_handler(IntentBuilder("ComputerOffIntent").require("Computer")
+                    .require("Off").optionally("Turn"))
     def handle_turn_off_intent(self, message):
         try:
             config = self.config_core.get("RemoteComputerSkill", {})
@@ -81,50 +94,63 @@ class RemoteComputerSkill(MycroftSkill):
                 user_password = str(self.settings.get("user_password"))
                 sudo_password = str(self.settings.get("sudo_password"))
 
-            try:
-                ip = ipaddress.ip_address(ip_address)
-
-            except ValueError:
-                self.speak_dialog("invalid", {"word": "I.P"})
-                return
-
-            if self.ask_yesno("ask.confirmation",
-                              {"word": "shut down"}) == "no":
-                self.speak_dialog("okay")
-                return
-
-            client = paramiko.SSHClient()
-            client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-            client.connect(
-                        hostname=str(ip),
-                        port=port,
-                        username=user,
-                        password=user_password)
-
-            transport = client.get_transport()
-
-            session = transport.open_session()
-            session.set_combine_stderr(True)
-            session.get_pty()
-
-            if sys.platform.startswith("win"):
-                session.exec_command("shutdown /s")
-            else:
-                session.exec_command("sudo -k shutdown -h now")
-
-            stdin = session.makefile('wb', -1)
-            stdout = session.makefile('rb', -1)
-            stdin.write(sudo_password + '\n')
-            stdin.flush()
-            stdout.read()
-
-            client.close()
-
-            self.speak_dialog("computer.off")
+            if not ip_address or not port or not user \
+                    or not user_password or not sudo_password:
+                raise Exception("None found.")
 
         except Exception as e:
-            self.speak_dialog("connection.error")
+            self.speak_dialog("settings.error")
             self.log.error(e)
+            return
+
+        try:
+            ip = ipaddress.ip_address(ip_address)
+
+        except ValueError:
+            self.speak_dialog("invalid", {"word": "I.P"})
+            return
+
+        prompt_response = self.ask_yesno("ask.confirmation",
+                                         {"word": "turn on"})
+
+        if prompt_response == "yes":
+            try:
+                client = paramiko.SSHClient()
+                client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+                client.connect(
+                            hostname=str(ip),
+                            port=port,
+                            username=user,
+                            password=user_password)
+
+                transport = client.get_transport()
+
+                session = transport.open_session()
+                session.set_combine_stderr(True)
+                session.get_pty()
+
+                if sys.platform.startswith("win"):
+                    session.exec_command("shutdown /s")
+
+                else:
+                    session.exec_command("sudo -k shutdown -h now")
+
+                stdin = session.makefile('wb', -1)
+                stdout = session.makefile('rb', -1)
+                stdin.write(sudo_password + '\n')
+                stdin.flush()
+                stdout.read()
+
+                client.close()
+
+                self.speak_dialog("computer.off")
+
+            except Exception as e:
+                self.speak_dialog("connection.error")
+                self.log.error(e)
+
+        elif prompt_response == "no":
+            self.speak_dialog("okay")
 
     def stop(self):
         pass
